+++ lib/src/extended_rich_text.dart
@@ -2,7 +2,6 @@ import 'dart:ui';
 import 'package:extended_text/src/extended_render_paragraph.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:extended_text_library/extended_text_library.dart';
 import 'text_overflow_widget.dart';
 
 ///  * [TextStyle], which discusses how to style text.
@@ -30,7 +29,6 @@ class ExtendedRichText extends MultiChildRenderObjectWidget {
     this.textScaleFactor = 1.0,
     this.maxLines,
     this.locale,
-    this.onSelectionChanged,
     this.strutStyle,
     this.textWidthBasis = TextWidthBasis.parent,
     this.selection,
@@ -77,8 +75,6 @@ class ExtendedRichText extends MultiChildRenderObjectWidget {
   final TextSelection? selection;
 
   final Color? selectionColor;
-
-  final TextSelectionChangedHandler? onSelectionChanged;
 
   /// The text to display in this widget.
   final InlineSpan text;
@@ -163,7 +159,6 @@ class ExtendedRichText extends MultiChildRenderObjectWidget {
       textWidthBasis: textWidthBasis,
       locale: locale ?? Localizations.maybeLocaleOf(context),
       selection: selection,
-      onSelectionChanged: onSelectionChanged,
       selectionColor: selectionColor,
       startHandleLayerLink: startHandleLayerLink,
       endHandleLayerLink: endHandleLayerLink,
@@ -193,7 +188,6 @@ class ExtendedRichText extends MultiChildRenderObjectWidget {
       ..locale = locale ?? Localizations.maybeLocaleOf(context)
       ..selection = selection
       ..selectionColor = selectionColor
-      ..onSelectionChanged = onSelectionChanged
       ..startHandleLayerLink = startHandleLayerLink
       ..endHandleLayerLink = endHandleLayerLink
       ..textHeightBehavior = textHeightBehavior